refactor: Split metrics_export.py into modular files (1763 → 4 files)
Split the monolithic metrics_export.py (1763 lines) into focused modules:
- metrics_common.py (301 lines): MetricDefinition, METRICS dict, helpers
- prometheus_exporter.py (951 lines): PrometheusExporter class + HTTP server
- influxdb_exporter.py (553 lines): InfluxDBExporter class
- metrics_export.py (96 lines): Thin re-export layer for backward compatibility

All files now well under 1500 line threshold. Existing imports continue
to work via the re-export layer.
